add miniatures module
