+++ js/main.js
@@ -1,3 +1,1 @@
-import { userPost } from './data.js';
-// eslint-disable-next-line no-console
-console.log(userPost);
+import './miniatures.js';
